requirements.txt
yt-dlp==2025.2.10
python-telegram-bot==13.15